Menu option#3
-> able to fetch data from txt file
-> able to select test case and display output

+++ src/Algorithm.java
@@ -4,7 +4,7 @@ public abstract class Algorithm {
 
 	protected ArrayList<Parcel> parcelList;
 	protected ArrayList<Truck> allocatedTrucks;
-	protected float timeAllocated;
+	protected long timeAllocated;
 	protected int totalRemainingCapacity;
 
 	public Algorithm() {
@@ -29,7 +29,7 @@ public abstract class Algorithm {
 		return this.totalRemainingCapacity;
 	}
 
-	protected void setTimeAllocated(float startTime, float endTime) {
+	protected void setTimeAllocated(long startTime, long endTime) {
 		this.timeAllocated = endTime - startTime;
 	}
